Fix ssr compatibility
