+++ packages/core/src/browser/timer.ts
@@ -1,12 +1,13 @@
 import { getZoneJsOriginalValue } from '../tools/getZoneJsOriginalValue'
 import { monitor } from '../tools/monitor'
+import { getGlobalObject } from '../tools/getGlobalObject'
 
 export type TimeoutId = ReturnType<typeof window.setTimeout>
 
 export function setTimeout(callback: () => void, delay?: number): TimeoutId {
-  return getZoneJsOriginalValue(window, 'setTimeout')(monitor(callback), delay)
+  return getZoneJsOriginalValue(getGlobalObject<Window>(), 'setTimeout')(monitor(callback), delay)
 }
 
 export function clearTimeout(timeoutId: TimeoutId | undefined) {
-  getZoneJsOriginalValue(window, 'clearTimeout')(timeoutId)
+  getZoneJsOriginalValue(getGlobalObject<Window>(), 'clearTimeout')(timeoutId)
 }
