+++ packages/core/src/index.ts
@@ -56,6 +56,7 @@ export * from './tools/display'
 export * from './tools/urlPolyfill'
 export * from './tools/timeUtils'
 export * from './tools/utils'
+export * from './tools/getGlobalObject'
 export * from './tools/createEventRateLimiter'
 export * from './tools/browserDetection'
 export { sendToExtension } from './tools/sendToExtension'
