+++ packages/core/src/tools/getZoneJsOriginalValue.ts
@@ -1,3 +1,5 @@
+import { getGlobalObject } from './getGlobalObject'
+
 export interface BrowserWindowWithZoneJs extends Window {
   Zone?: {
     // All Zone.js versions expose the __symbol__ method, but we observed that some website have a
@@ -24,7 +26,7 @@ export function getZoneJsOriginalValue<Target, Name extends keyof Target & strin
   target: Target,
   name: Name
 ): Target[Name] {
-  const browserWindow = window as BrowserWindowWithZoneJs
+  const browserWindow = getGlobalObject<BrowserWindowWithZoneJs>()
   let original: Target[Name] | undefined
   if (browserWindow.Zone && typeof browserWindow.Zone.__symbol__ === 'function') {
     original = (target as any)[browserWindow.Zone.__symbol__(name)]
